Reduce default buffer frames and add result saving feature
Updated `main.cpp` to change the default number of buffer pool frames from 10 to 5. Introduced a new feature that prompts the user to decide whether to save query results to a file. If the user opts to save, they can specify an output file name, which will automatically have a ".txt" extension if not provided. Added a vector to store all results and implemented functionality to write these results to the specified file, along with a success message indicating the number of records saved.